mod: Refactor diagnosis response data structure
Separate symptomitem, drugitem data model for other class

Related to: #28
