+++ client/data/src/main/java/com/peekaboo/data/entity/response/dianosis/SymptomItem.kt
@@ -1,0 +1,10 @@
+package com.peekaboo.data.entity.response.dianosis
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class SymptomItem(
+    @SerialName("name")
+    val name: String = "",
+)
