+++ client/domain/src/main/java/com/peekaboo/domain/entity/response/diagnosis/DrugItem.kt
@@ -1,0 +1,6 @@
+package com.peekaboo.domain.entity.response.diagnosis
+
+data class DrugItem(
+    val name: String = "",
+    val efficacy: String = "",
+)
