+++ client/domain/src/main/java/com/peekaboo/domain/entity/response/diagnosis/SymptomItem.kt
@@ -1,0 +1,5 @@
+package com.peekaboo.domain.entity.response.diagnosis
+
+data class SymptomItem(
+    val name: String = "",
+)
